refactor: move activation functions to activations module
